+++ snd-dac.c
@@ -656,7 +656,7 @@ static void stop_playing_with_toggle(dac_info *dp, dac_toggle_t toggle, with_hoo
 	  (sp->inuse == SOUND_NORMAL) && 
 	  (sp->index >= 0))
 	{
-	  int i;
+	  unsigned int i;
 	  for (i = 0; i < sp->nchans; i++)
 	    {
 	      chan_info *cp;
@@ -1354,7 +1354,7 @@ static dac_info *play_sound_1(snd_info *sp, mus_long_t start, mus_long_t end, pl
 			      Xen edpos, Xen stop_proc, const char *caller, int arg_pos)
 {
   /* just plays one sound (ignores possible sync) */
-  int i;
+  unsigned int i;
   dac_info *dp = NULL, *rtn_dp = NULL;
 
   if ((background == NOT_IN_BACKGROUND) && 
@@ -2614,7 +2614,7 @@ void clear_players(void)
   int i;
   for (i = 0; i < players_size; i++)
     {
-      int j;
+      unsigned int j;
       snd_info *sp;
       sp = players[i];
       if (sp)
@@ -3018,7 +3018,7 @@ If object is a string, it is assumed to be a file name: \n    " play_example "\n
     play_sound_1(sp, start, end, background, edit_position, stop_func, S_play, edpos_argpos);
   else 
     {
-      if ((channel < sp->nchans) &&
+      if ((channel <(int)(sp->nchans)) &&
 	  (channel >= 0))
 	{
 	  int pos;
@@ -3118,7 +3118,7 @@ static Xen g_player_home(Xen player)
       (index < players_size) && 
       (players[index]) &&
       (players[index]->chans) &&
-      (player_chans[index] < players[index]->nchans))
+      (player_chans[index] < (int)players[index]->nchans))
     {
       chan_info *cp;
       cp = players[index]->chans[player_chans[index]]; /* trying to get back to the original sound index (not the player index) */
